third_party/blink/web_tests/platform/generic/fast/loader/page-dismissal-modal-dialogs-expected.txt

CONSOLE ERROR: Blocked alert('mainFrame beforeunload PASS') during beforeunload.
CONSOLE ERROR: Blocked confirm('mainFrame beforeunload PASS') during beforeunload.
CONSOLE ERROR: Blocked prompt('mainFrame beforeunload PASS') during beforeunload.
CONSOLE ERROR: Blocked alert('iFrame click PASS') during beforeunload.
CONSOLE ERROR: Blocked confirm('iFrame click PASS') during beforeunload.
CONSOLE ERROR: Blocked prompt('iFrame click PASS') during beforeunload.
CONSOLE ERROR: Blocked alert('iFrame beforeunload PASS') during beforeunload.
CONSOLE ERROR: Blocked confirm('iFrame beforeunload PASS') during beforeunload.
CONSOLE ERROR: Blocked prompt('iFrame beforeunload PASS') during beforeunload.
CONSOLE ERROR: Blocked alert('mainFrame click PASS') during beforeunload.
CONSOLE ERROR: Blocked confirm('mainFrame click PASS') during beforeunload.
CONSOLE ERROR: Blocked prompt('mainFrame click PASS') during beforeunload.
CONSOLE ERROR: Blocked alert('mainFrame pagehide PASS') during pagehide.
CONSOLE ERROR: Blocked confirm('mainFrame pagehide PASS') during pagehide.
CONSOLE ERROR: Blocked prompt('mainFrame pagehide PASS') during pagehide.
CONSOLE ERROR: Blocked alert('iFrame click PASS') during pagehide.
CONSOLE ERROR: Blocked confirm('iFrame click PASS') during pagehide.
CONSOLE ERROR: Blocked prompt('iFrame click PASS') during pagehide.
CONSOLE ERROR: Blocked alert('mainFrame unload PASS') during unload.
CONSOLE ERROR: Blocked confirm('mainFrame unload PASS') during unload.
CONSOLE ERROR: Blocked prompt('mainFrame unload PASS') during unload.
CONSOLE ERROR: Blocked alert('iFrame click PASS') during unload.
CONSOLE ERROR: Blocked confirm('iFrame click PASS') during unload.
CONSOLE ERROR: Blocked prompt('iFrame click PASS') during unload.
CONSOLE ERROR: Blocked alert('iFrame pagehide PASS') during pagehide.
CONSOLE ERROR: Blocked confirm('iFrame pagehide PASS') during pagehide.
CONSOLE ERROR: Blocked prompt('iFrame pagehide PASS') during pagehide.
CONSOLE ERROR: Blocked alert('mainFrame click PASS') during pagehide.
CONSOLE ERROR: Blocked confirm('mainFrame click PASS') during pagehide.
CONSOLE ERROR: Blocked prompt('mainFrame click PASS') during pagehide.
CONSOLE ERROR: Blocked alert('iFrame unload PASS') during unload.
CONSOLE ERROR: Blocked confirm('iFrame unload PASS') during unload.
CONSOLE ERROR: Blocked prompt('iFrame unload PASS') during unload.
CONSOLE ERROR: Blocked alert('mainFrame click PASS') during unload.
CONSOLE ERROR: Blocked confirm('mainFrame click PASS') during unload.
CONSOLE ERROR: Blocked prompt('mainFrame click PASS') during unload.
PASS
